Corrected mistakes
